requirements.txt
discord
pyyaml